requirements.txt
discord
pyyaml